moved package to correct path, added formatter tests
