--- code/OpenDiabetes-NSApi/src/de/opendiabetes/nsapi/exception/NightscoutDataException.java
@@ -1,14 +1,0 @@
-package de.opendiabetes.nsapi.exception;
-
-/**
- * Representing that invalid data was provided.
- */
-public class NightscoutDataException extends RuntimeException {
-    public NightscoutDataException(String message) {
-        super(message);
-    }
-
-    public NightscoutDataException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
